ms_proveedores

+++ src/main/java/cl/kiosko/ms_proveedores/service/ProveedorService.java
@@ -3,14 +3,12 @@ package cl.kiosko.ms_proveedores.service;
 import cl.kiosko.ms_proveedores.DTO.ProveedorDTO;
 import cl.kiosko.ms_proveedores.model.Proveedor;
 import cl.kiosko.ms_proveedores.repository.ProveedorRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
 public class ProveedorService {
-    @Autowired
 
     private final ProveedorRepository repo;
 
@@ -40,7 +38,24 @@ public class ProveedorService {
         return repo.save(proveedor);
     }
 
+    public Proveedor actualizar(Long id, ProveedorDTO dto) {
+        Proveedor proveedor = repo.findById(id)
+                .orElseThrow(() -> new RuntimeException("Proveedor no encontrado"));
+
+        proveedor.setNombreEmpresa(dto.getNombreEmpresa());
+        proveedor.setRut(dto.getRut());
+        proveedor.setDireccion(dto.getDireccion());
+        proveedor.setTelefono(dto.getTelefono());
+        proveedor.setEmail(dto.getEmail());
+        proveedor.setProducto(dto.getProducto());
+
+        return repo.save(proveedor);
+    }
+
     public void eliminar(Long id) {
+        if (!repo.existsById(id)) {
+            throw new RuntimeException("Proveedor no existe");
+        }
         repo.deleteById(id);
     }
 }
